Patched bug where command could be sent to closed SPJS
Edited the update method so that it displays long verify time comments
with commas in them to make it easier to differentiate thousands of
milliseconds.
Updated some of the publish commands to use new message sending methods.
Patched bug where the spjsSend method would allow list commands to be
sent to the SPJS while it is not open and would not flag the commands as
errored correctly, leading to the inability to send list commands until
the command stale limit was hit.